Updates the ACR statement to a SHOULD for the baseline
This weakens the ACR statements from what the CDS currently prescribes because it is currently unclear
whether supporting the ACR claim in each DH actually provides value in the CDR Action Intiation world.

FAPI does not require it, and the future state for RAR request details is that each of them will require
a minimum ACR anyway so the ADR doesn't have a choice.

The only real reason for enforcing holders to support the claim is to allow the ADR to request an ACR
higher than what their request requires, but I feel like that violates the minimum friction CDR principle.

We can add it as a SHALL in CDR Sharing V1